Re-arrange menus initialization
Every menu is responsible for its layout and load its sons. Only MainMenu has
a contructor in which set the shared members (i.e. the static member of Menu)
mConfigManager, mSettinsManager and CEGUISystem. The others have a constructor
in which is set the father

Declare the callback to cancel button in SettingsMenu

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>