Fix thread usage
Ogre requires that the main, and only the main thread access render functions.
So we changed the mechanism: now it's created a thread to run the Game::run
function (i.e. the game's logic) and the main thread calls UI::start_thread,
which starts the rendering system. Actually, it waits for a call to UI::start
to start rendering. The later is called by Game::run when the board and
players are ready.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>