Change UI::gameOver call
This change is made so the GUI can show the user who is ther winner. The new
call is UI::gameOver(Move const& start, Move const& direction), where start is
the position where to start adding direction, defining therefore a line inside
the cube where a player won.

As a good "side-effect", in UI3d when game ends it returns to GUI.

Signed-off-by: Lucas De Marchi <lucas.de.marchi@gmail.com>